requirements.txt
amqp==5.1.1
anyio==3.6.2
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.6
Django==4.1.7
django-celery-beat==2.4.0
django-rest-framework==0.1.0
django-timezone-field==5.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.25.1
flake8==6.0.0
flake8-quotes==3.3.2
flake8-variables-names==0.0.5
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==0.16.3
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
kombu==5.2.4
mccabe==0.7.0
pep8-naming==0.13.3
prompt-toolkit==3.0.38
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
PyJWT==2.6.0
pyrsistent==0.19.3
pyTelegramBotAPI==4.10.0
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.1
pytz==2022.7.1
PyYAML==6.0
redis==4.5.1
requests==2.28.2
rfc3986==1.5.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
stripe==5.2.0
telebot==0.0.5
telegram==0.0.1
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
